adds all basics (untested) of convolution_backward_device
This is a significant commit with many changes in the
dev_layer_conv.cu file which relate to the
convolution_backward_device.  Although the tests were
not run, and are likely breaking.  This commit includes
a (at least partially) running function.

+++ include/awnn/common.h
@@ -19,7 +19,7 @@ extern "C" {
 #endif
 
 typedef unsigned int uint;
-typedef double T;
+typedef float T;
 
 typedef unsigned int label_t;
 label_t *label_make_random(uint nr_elem, uint range);
